+++ tests/Linna/TypedArrayObject/ArrayOfBooleansTest.php
@@ -9,23 +9,22 @@
  */
 declare(strict_types=1);
 
-namespace Linna\Tests;
+namespace Linna\TypedArrayObject;
 
 use InvalidArgumentException;
-use Linna\TypedArrayObject\StringArrayObject;
 use PHPUnit\Framework\TestCase;
 
 /**
- * String Array Object Test.
+ * Bool Array Object Test.
  */
-class StringArrayObjectTest extends TestCase
+class ArrayOfBooleansTest extends TestCase
 {
     /**
      * Test new instance.
      */
     public function testNewInstance(): void
     {
-        $this->assertInstanceOf(StringArrayObject::class, (new StringArrayObject()));
+        $this->assertInstanceOf(ArrayOfBooleans::class, (new ArrayOfBooleans()));
     }
 
     /**
@@ -33,7 +32,7 @@ class StringArrayObjectTest extends TestCase
      */
     public function testNewInstanceWithValidArgument(): void
     {
-        $this->assertInstanceOf(StringArrayObject::class, (new StringArrayObject(['a','b','c','d','e','f','g','h','i'])));
+        $this->assertInstanceOf(ArrayOfBooleans::class, (new ArrayOfBooleans([true,true,true,false])));
     }
 
     /**
@@ -41,25 +40,25 @@ class StringArrayObjectTest extends TestCase
      */
     public function testSetValueWithValidArgument(): void
     {
-        $stringArray = new StringArrayObject();
-        $stringArray[] = 'a';
-        
-        $this->assertSame(1, $this->count($stringArray));
-        $this->assertSame('a', $stringArray[0]);
+        $boolArray = new ArrayOfBooleans();
+        $boolArray[] = false;
+
+        $this->assertSame(1, $this->count($boolArray));
+        $this->assertSame(false, $boolArray[0]);
     }
-    
+
     /**
      * Test append value with valid argument.
      */
     public function testAppendValueWithValidArgument(): void
     {
-        $stringArray = new StringArrayObject();
-        $stringArray->append('a');
-        
-        $this->assertSame(1, $this->count($stringArray));
-        $this->assertSame('a', $stringArray[0]);
+        $boolArray = new ArrayOfBooleans();
+        $boolArray->append(false);
+
+        $this->assertSame(1, $this->count($boolArray));
+        $this->assertSame(false, $boolArray[0]);
     }
-    
+
     /**
      * Provide invalid typed arrays.
      *
@@ -69,25 +68,26 @@ class StringArrayObjectTest extends TestCase
     {
         return [
             [[[1], [2]]], //array
-            [[true, false]], //bool
-            [[function () {}, function () {}]], //callable
+            [[function () {
+            }, function () {
+            }]], //callable
             [[1.1, 2.2]], //float
             [[1, 2]], //int
             [[(object) ['name' => 'foo'], (object) ['name' => 'bar']]], //object
-            //[['a', 'b']], //string
+            [['a', 'b']], //string
         ];
     }
 
     /**
      * Test new instance with invalid argument.
-     * 
+     *
      * @dataProvider invalidArrayProvider
      */
     public function testNewInstanceWithInvalidArgument(array $array): void
     {
         $this->expectException(InvalidArgumentException::class);
-        
-        $stringArray = new StringArrayObject($array);
+
+        (new ArrayOfBooleans($array));
     }
 
     /**
@@ -99,38 +99,37 @@ class StringArrayObjectTest extends TestCase
     {
         return [
             [[1]], //array
-            [true], //bool
-            [function () {}], //callable
+            [function () {
+            }], //callable
             [1.1], //float
             [1], //int
             [(object) ['name' => 'foo']], //object
-            //['a'], //string
+            ['a'], //string
         ];
     }
 
     /**
      * Test set value with invalid argument.
-     * 
+     *
      * @dataProvider invalidValueProvider
      */
     public function testSetValueWithInvalidArgument($value): void
     {
         $this->expectException(InvalidArgumentException::class);
-     
-        $stringArray = new StringArrayObject();
-        $stringArray[] = $value;
+
+        $boolArray = new ArrayOfBooleans();
+        $boolArray[] = $value;
     }
 
     /**
      * Test append value with invalid argument.
-     * 
+     *
      * @dataProvider invalidValueProvider
      */
     public function testAppendValueWithInvalidArgument($value): void
     {
         $this->expectException(InvalidArgumentException::class);
-        
-        $stringArray = new StringArrayObject();
-        $stringArray->append($value);
+
+        (new ArrayOfBooleans())->append($value);
     }
 }