+++ tests/Linna/TypedArrayObject/ArrayOfFloatsTest.php
@@ -9,23 +9,22 @@
  */
 declare(strict_types=1);
 
-namespace Linna\Tests;
+namespace Linna\TypedArrayObject;
 
 use InvalidArgumentException;
-use Linna\TypedArrayObject\FloatArrayObject;
 use PHPUnit\Framework\TestCase;
 
 /**
  * Int Array Object Test.
  */
-class FloatArrayObjectTest extends TestCase
+class ArrayOfFloatsTest extends TestCase
 {
     /**
      * Test new instance.
      */
     public function testNewInstance(): void
     {
-        $this->assertInstanceOf(FloatArrayObject::class, (new FloatArrayObject()));
+        $this->assertInstanceOf(ArrayOfFloats::class, (new ArrayOfFloats()));
     }
 
     /**
@@ -33,7 +32,7 @@ class FloatArrayObjectTest extends TestCase
      */
     public function testNewInstanceWithValidArgument(): void
     {
-        $this->assertInstanceOf(FloatArrayObject::class, (new FloatArrayObject([1.1,2.2,3.3,4.4,5.5])));
+        $this->assertInstanceOf(ArrayOfFloats::class, (new ArrayOfFloats([1.1,2.2,3.3,4.4,5.5])));
     }
 
     /**
@@ -41,25 +40,25 @@ class FloatArrayObjectTest extends TestCase
      */
     public function testSetValueWithValidArgument(): void
     {
-        $floatArray = new FloatArrayObject();
+        $floatArray = new ArrayOfFloats();
         $floatArray[] = 1.1;
-        
+
         $this->assertSame(1, $this->count($floatArray));
         $this->assertSame(1.1, $floatArray[0]);
     }
-    
+
     /**
      * Test append value with valid argument.
      */
     public function testAppendValueWithValidArgument(): void
     {
-        $floatArray = new FloatArrayObject();
+        $floatArray = new ArrayOfFloats();
         $floatArray->append(1.1);
-        
+
         $this->assertSame(1, $this->count($floatArray));
         $this->assertSame(1.1, $floatArray[0]);
     }
-    
+
     /**
      * Provide invalid typed arrays.
      *
@@ -70,8 +69,9 @@ class FloatArrayObjectTest extends TestCase
         return [
             [[[1], [2]]], //array
             [[true, false]], //bool
-            [[function () {}, function () {}]], //callable
-            //[[1.1, 2.2]], //float
+            [[function () {
+            }, function () {
+            }]], //callable
             [[1, 2]], //int
             [[(object) ['name' => 'foo'], (object) ['name' => 'bar']]], //object
             [['a', 'b']], //string
@@ -80,14 +80,14 @@ class FloatArrayObjectTest extends TestCase
 
     /**
      * Test new instance with invalid argument.
-     * 
+     *
      * @dataProvider invalidArrayProvider
      */
     public function testNewInstanceWithInvalidArgument(array $array): void
     {
         $this->expectException(InvalidArgumentException::class);
-        
-        $floatArray = new FloatArrayObject($array);
+
+        (new ArrayOfFloats($array));
     }
 
     /**
@@ -100,8 +100,8 @@ class FloatArrayObjectTest extends TestCase
         return [
             [[1]], //array
             [true], //bool
-            [function () {}], //callable
-            //[1.1], //float
+            [function () {
+            }], //callable
             [1], //int
             [(object) ['name' => 'foo']], //object
             ['a'], //string
@@ -110,27 +110,26 @@ class FloatArrayObjectTest extends TestCase
 
     /**
      * Test set value with invalid argument.
-     * 
+     *
      * @dataProvider invalidValueProvider
      */
     public function testSetValueWithInvalidArgument($value): void
     {
         $this->expectException(InvalidArgumentException::class);
-     
-        $floatArray = new FloatArrayObject();
+
+        $floatArray = new ArrayOfFloats();
         $floatArray[] = $value;
     }
 
     /**
      * Test append value with invalid argument.
-     * 
+     *
      * @dataProvider invalidValueProvider
      */
     public function testAppendValueWithInvalidArgument($value): void
     {
         $this->expectException(InvalidArgumentException::class);
-        
-        $floatArray = new FloatArrayObject();
-        $floatArray->append($value);
+
+        (new ArrayOfFloats())->append($value);
     }
 }