+++ tests/Linna/TypedArrayObject/ArrayOfObjectsTest.php
@@ -9,23 +9,22 @@
  */
 declare(strict_types=1);
 
-namespace Linna\Tests;
+namespace Linna\TypedArrayObject;
 
 use InvalidArgumentException;
-use Linna\TypedArrayObject\IntArrayObject;
 use PHPUnit\Framework\TestCase;
 
 /**
  * Int Array Object Test.
  */
-class IntArrayObjectTest extends TestCase
+class ArrayOfObjectsTest extends TestCase
 {
     /**
      * Test new instance.
      */
     public function testNewInstance(): void
     {
-        $this->assertInstanceOf(IntArrayObject::class, (new IntArrayObject()));
+        $this->assertInstanceOf(ArrayOfObjects::class, (new ArrayOfObjects()));
     }
 
     /**
@@ -33,7 +32,7 @@ class IntArrayObjectTest extends TestCase
      */
     public function testNewInstanceWithValidArgument(): void
     {
-        $this->assertInstanceOf(IntArrayObject::class, (new IntArrayObject([1,2,3,4,5,6,7,8,9,0])));
+        $this->assertInstanceOf(ArrayOfObjects::class, (new ArrayOfObjects([(object)["name" => "test"]])));
     }
 
     /**
@@ -41,25 +40,29 @@ class IntArrayObjectTest extends TestCase
      */
     public function testSetValueWithValidArgument(): void
     {
-        $intArray = new IntArrayObject();
-        $intArray[] = 1;
-        
-        $this->assertSame(1, $this->count($intArray));
-        $this->assertSame(1, $intArray[0]);
+        $object = (object) ["foo" => 1, "bar" => 2,"baz" => 3];
+
+        $objectArray = new ArrayOfObjects();
+        $objectArray[] = $object;
+
+        $this->assertSame(1, $this->count($objectArray));
+        $this->assertSame($object, $objectArray[0]);
     }
-    
+
     /**
      * Test append value with valid argument.
      */
     public function testAppendValueWithValidArgument(): void
     {
-        $intArray = new IntArrayObject();
-        $intArray->append(1);
-        
-        $this->assertSame(1, $this->count($intArray));
-        $this->assertSame(1, $intArray[0]);
+        $object = (object) ["foo" => 1, "bar" => 2,"baz" => 3];
+
+        $objectArray = new ArrayOfObjects();
+        $objectArray->append($object);
+
+        $this->assertSame(1, $this->count($objectArray));
+        $this->assertSame($object, $objectArray[0]);
     }
-    
+
     /**
      * Provide invalid typed arrays.
      *
@@ -70,24 +73,22 @@ class IntArrayObjectTest extends TestCase
         return [
             [[[1], [2]]], //array
             [[true, false]], //bool
-            [[function () {}, function () {}]], //callable
             [[1.1, 2.2]], //float
-            //[[1, 2]], //int
-            [[(object) ['name' => 'foo'], (object) ['name' => 'bar']]], //object
+            [[1, 2]], //int
             [['a', 'b']], //string
         ];
     }
 
     /**
      * Test new instance with invalid argument.
-     * 
+     *
      * @dataProvider invalidArrayProvider
      */
     public function testNewInstanceWithInvalidArgument(array $array): void
     {
         $this->expectException(InvalidArgumentException::class);
-        
-        $intArray = new IntArrayObject($array);
+
+        (new ArrayOfObjects($array));
     }
 
     /**
@@ -100,37 +101,34 @@ class IntArrayObjectTest extends TestCase
         return [
             [[1]], //array
             [true], //bool
-            [function () {}], //callable
             [1.1], //float
-            //[1], //int
-            [(object) ['name' => 'foo']], //object
+            [1], //int
             ['a'], //string
         ];
     }
 
     /**
      * Test set value with invalid argument.
-     * 
+     *
      * @dataProvider invalidValueProvider
      */
     public function testSetValueWithInvalidArgument($value): void
     {
         $this->expectException(InvalidArgumentException::class);
-     
-        $intArray = new IntArrayObject();
-        $intArray[] = $value;
+
+        $objectArray = new ArrayOfObjects();
+        $objectArray[] = $value;
     }
 
     /**
      * Test append value with invalid argument.
-     * 
+     *
      * @dataProvider invalidValueProvider
      */
     public function testAppendValueWithInvalidArgument($value): void
     {
         $this->expectException(InvalidArgumentException::class);
-        
-        $intArray = new IntArrayObject();
-        $intArray->append($value);
+
+        (new ArrayOfObjects())->append($value);
     }
 }